+++ tw-entrypoints/src/main/java/com/transferwise/common/entrypoints/tableaccessstatistics/DefaultTasSqlFilter.java
@@ -1,0 +1,14 @@
+package com.transferwise.common.entrypoints.tableaccessstatistics;
+
+import org.apache.commons.lang3.StringUtils;
+
+public class DefaultTasSqlFilter implements TasSqlFilter {
+
+  @Override
+  public boolean skip(String sql) {
+    if (StringUtils.startsWithIgnoreCase(sql, "SET statement_timeout TO")) {
+      return true;
+    }
+    return false;
+  }
+}