+++ tw-entrypoints/src/main/java/com/transferwise/common/entrypoints/tableaccessstatistics/ParsedQuery.java
@@ -1,0 +1,32 @@
+package com.transferwise.common.entrypoints.tableaccessstatistics;
+
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+import lombok.Data;
+import lombok.experimental.Accessors;
+
+@Data
+@Accessors(chain = true)
+public class ParsedQuery {
+
+  private Map<String, SqlOperation> operations = new HashMap<>();
+
+  ParsedQuery addOperation(String operationName, SqlOperation operation) {
+    operations.put(operationName, operation);
+    return this;
+  }
+
+  @Data
+  @Accessors(chain = true)
+  static class SqlOperation {
+
+    private Set<String> tableNames = new HashSet<>();
+
+    public SqlOperation addTable(String table) {
+      tableNames.add(table);
+      return this;
+    }
+  }
+}