+++ tw-entrypoints/src/main/java/com/transferwise/common/entrypoints/tableaccessstatistics/TableAccessStatisticsParsedQueryRegistry.java
@@ -1,0 +1,11 @@
+package com.transferwise.common.entrypoints.tableaccessstatistics;
+
+/**
+ * Allows to register queries which are not parseable or very slow to parse.
+ */
+public interface TableAccessStatisticsParsedQueryRegistry {
+
+  void register(String sql, ParsedQuery parsedQuery);
+
+  ParsedQuery get(String sql);
+}