+++ tw-entrypoints/src/main/java/com/transferwise/common/entrypoints/tableaccessstatistics/TableAccessStatisticsSpyqlListener.java
@@ -12,7 +12,8 @@ import com.transferwise.common.baseutils.meters.cache.TagsSet;
 import com.transferwise.common.context.TwContext;
 import com.transferwise.common.context.TwContextMetricsTemplate;
 import com.transferwise.common.entrypoints.EntryPointsMetrics;
-import com.transferwise.common.entrypoints.tableaccessstatistics.TableAccessStatisticsSpyqlListener.SqlParseResult.SqlOperation;
+import com.transferwise.common.entrypoints.EntryPointsProperties;
+import com.transferwise.common.entrypoints.tableaccessstatistics.ParsedQuery.SqlOperation;
 import com.transferwise.common.spyql.event.GetConnectionEvent;
 import com.transferwise.common.spyql.event.StatementExecuteEvent;
 import com.transferwise.common.spyql.event.StatementExecuteFailureEvent;
@@ -21,20 +22,12 @@ import com.transferwise.common.spyql.listener.SpyqlDataSourceListener;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tag;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Set;
-import java.util.concurrent.Executor;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
-import lombok.Data;
-import lombok.NonNull;
-import lombok.experimental.Accessors;
 import lombok.extern.slf4j.Slf4j;
 import net.sf.jsqlparser.statement.Statement;
-import net.sf.jsqlparser.statement.Statements;
 import org.apache.commons.lang3.StringUtils;
 
 @Slf4j
@@ -46,6 +39,7 @@ public class TableAccessStatisticsSpyqlListener implements SpyqlDataSourceListen
 
   public static final String COUNTER_PARSES = "EntryPoints_Tas_Parses";
   public static final String COUNTER_FAILED_PARSES = "EntryPoints_Tas_FailedParses";
+  public static final String COUNTER_SLOW_PARSES = "EntryPoints_Tas_SlowParses";
   public static final String COUNTER_UNCOUNTED_QUERIES = "EntryPoints_Tas_UncountedQueries";
   public static final String TIMER_FIRST_TABLE_ACCESS = "EntryPoints_Tas_FirstTableAccess";
   public static final String COUNTER_TABLE_ACCESS = "EntryPoints_Tas_TableAccess";
@@ -63,19 +57,30 @@ public class TableAccessStatisticsSpyqlListener implements SpyqlDataSourceListen
 
   private final IMeterCache meterCache;
 
-  final LoadingCache<String, SqlParseResult> sqlParseResultsCache;
+  final LoadingCache<String, ParsedQuery> sqlParseResultsCache;
 
-  public TableAccessStatisticsSpyqlListener(IMeterCache meterCache, Executor executor, String databaseName,
-      long sqlParserCacheSizeMib) {
+  private final TableAccessStatisticsParsedQueryRegistry tableAccessStatisticsParsedQueryRegistry;
+  private final SqlParser sqlParser;
+  private final EntryPointsProperties entryPointsProperties;
+  private final TasSqlFilter sqlFilter;
+
+  public TableAccessStatisticsSpyqlListener(IMeterCache meterCache, ExecutorService executorService,
+      TableAccessStatisticsParsedQueryRegistry tableAccessStatisticsParsedQueryRegistry, String databaseName,
+      EntryPointsProperties entryPointsProperties, TasSqlFilter sqlFilter) {
     this.databaseName = databaseName;
     this.dbTag = Tag.of(EntryPointsMetrics.TAG_DATABASE, databaseName);
     this.meterCache = meterCache;
+    this.tableAccessStatisticsParsedQueryRegistry = tableAccessStatisticsParsedQueryRegistry;
+    this.sqlParser = new SqlParser(executorService);
+    this.entryPointsProperties = entryPointsProperties;
+    this.sqlFilter = sqlFilter;
+
     MeterRegistry meterRegistry = meterCache.getMeterRegistry();
     meterRegistry.config().meterFilter(new TasMeterFilter());
 
-    sqlParseResultsCache = Caffeine.newBuilder().maximumWeight(sqlParserCacheSizeMib * MIB).recordStats()
-        .executor(executor)
-        .weigher((String k, SqlParseResult v) -> k.length() * 2)
+    sqlParseResultsCache = Caffeine.newBuilder().maximumWeight(entryPointsProperties.getTas().getSqlParser().getCacheSizeMib() * MIB).recordStats()
+        .executor(executorService)
+        .weigher((String k, ParsedQuery v) -> k.length() * 2)
         .build(sql -> parseSql(sql, TwContext.current()));
 
     Gauge.builder(GAUGE_SQL_PARSER_RESULT_CACHE_SIZE, sqlParseResultsCache::estimatedSize).register(meterRegistry);
@@ -88,26 +93,41 @@ public class TableAccessStatisticsSpyqlListener implements SpyqlDataSourceListen
     return new ConnectionListener();
   }
 
-  protected SqlParseResult parseSql(String sql, TwContext context) {
-    SqlParseResult result = new SqlParseResult();
+  protected ParsedQuery parseSql(String sql, TwContext context) {
+    ParsedQuery result = new ParsedQuery();
+    long startTimeMs = System.currentTimeMillis();
     try {
-      Statements stmts = SqlParserUtils.parseToStatements(sql);
+      if (sqlFilter.skip(sql)) {
+        return result;
+      }
+
+      var stmts = sqlParser.parse(sql, entryPointsProperties.getTas().getSqlParser().getTimeout());
 
       for (Statement stmt : stmts.getStatements()) {
         // Intern() makes later equal checks much faster.
         String opName = getOperationName(stmt).intern();
         CustomTablesNamesFinder tablesNamesFinder = new CustomTablesNamesFinder();
-        List<String> tableNames = tablesNamesFinder.getTableList(stmt);
+        List<String> tableNames = null;
+        try {
+          tableNames = tablesNamesFinder.getTableList(stmt);
+        } catch (UnsupportedOperationException e) {
+          // Some type of statements do not support finding table names.
+          // For example a statement 'SHOW FULL TABLES IN ...'.
+          log.debug("Unsupported query '{}'.", sql, e);
+        }
 
-        SqlParseResult.SqlOperation sqlOp = result
+        ParsedQuery.SqlOperation sqlOp = result
             .getOperations()
-            .computeIfAbsent(opName, k -> new SqlParseResult.SqlOperation());
+            .computeIfAbsent(opName, k -> new ParsedQuery.SqlOperation());
 
-        for (String tableName : tableNames) {
-          // Intern() makes later equal checks much faster.
-          sqlOp.getTableNames().add(tableName.intern());
+        if (tableNames != null) {
+          for (String tableName : tableNames) {
+            // Intern() makes later equal checks much faster.
+            sqlOp.getTableNames().add(tableName.intern());
+          }
         }
       }
+
       meterCache.counter(COUNTER_PARSES, TagsSet.of(
           EntryPointsMetrics.TAG_DATABASE, databaseName,
           TwContextMetricsTemplate.TAG_EP_GROUP, context.getGroup(),
@@ -121,7 +141,24 @@ public class TableAccessStatisticsSpyqlListener implements SpyqlDataSourceListen
           TwContextMetricsTemplate.TAG_EP_NAME, context.getName(),
           TwContextMetricsTemplate.TAG_EP_OWNER, context.getOwner()
       )).increment();
-      log.debug(t.getMessage(), t);
+
+      // If we log this as error or even as info, we will already create spam for many Flyway queries.
+      log.debug("Parsing statement '{}' failed. You can use `TableAccessStatisticsParsedQueryRegistry` to register the query parse result manually.",
+          sql, t);
+    } finally {
+      long durationMs = System.currentTimeMillis() - startTimeMs;
+      if (durationMs > entryPointsProperties.getTas().getSqlParser().getParseDurationWarnThreshold().toMillis()) {
+        log.warn(
+            "Statement '{}' parsing took {} ms. You may want to use `TableAccessStatisticsParsedQueryRegistry` to register the"
+                + " query parse result manually. ");
+
+        meterCache.counter(COUNTER_SLOW_PARSES, TagsSet.of(
+            EntryPointsMetrics.TAG_DATABASE, databaseName,
+            TwContextMetricsTemplate.TAG_EP_GROUP, context.getGroup(),
+            TwContextMetricsTemplate.TAG_EP_NAME, context.getName(),
+            TwContextMetricsTemplate.TAG_EP_OWNER, context.getOwner()
+        )).increment();
+      }
     }
 
     return result;
@@ -149,9 +186,12 @@ public class TableAccessStatisticsSpyqlListener implements SpyqlDataSourceListen
       final Tag inTransactionTag = isInTransaction ? TAG_IN_TRANSACTION_TRUE : TAG_IN_TRANSACTION_FALSE;
       final Tag successTag = succeeded ? TAG_SUCCESS_TRUE : TAG_SUCCESS_FALSE;
 
-      SqlParseResult sqlParseResult = sqlParseResultsCache.get(sql, sqlForCache -> parseSql(sqlForCache, context));
+      var parsedQuery = tableAccessStatisticsParsedQueryRegistry.get(sql);
+      if (parsedQuery == null) {
+        parsedQuery = sqlParseResultsCache.get(sql, sqlForCache -> parseSql(sqlForCache, context));
+      }
 
-      if (sqlParseResult == null) {
+      if (parsedQuery == null) {
         meterCache.counter(COUNTER_UNCOUNTED_QUERIES, TagsSet.of(
             EntryPointsMetrics.TAG_DATABASE, databaseName,
             TwContextMetricsTemplate.TAG_EP_GROUP, context.getGroup(),
@@ -162,43 +202,30 @@ public class TableAccessStatisticsSpyqlListener implements SpyqlDataSourceListen
         return;
       }
 
-      for (Entry<String, SqlOperation> entry : sqlParseResult.operations.entrySet()) {
+      for (Entry<String, SqlOperation> entry : parsedQuery.getOperations().entrySet()) {
         String opName = entry.getKey();
         SqlOperation op = entry.getValue();
-        String firstTableName = null;
-        for (String tableName : op.getTableNames()) {
-          TagsSet tagsSet = TagsSet.of(
-              dbTag.getKey(), dbTag.getValue(),
-              TwContextMetricsTemplate.TAG_EP_GROUP, context.getGroup(),
-              TwContextMetricsTemplate.TAG_EP_NAME, context.getName(),
-              TwContextMetricsTemplate.TAG_EP_OWNER, context.getOwner(),
-              inTransactionTag.getKey(), inTransactionTag.getValue(),
-              TAG_OPERATION, opName,
-              successTag.getKey(), successTag.getValue(),
-              TAG_TABLE, tableName);
+        if (op.getTableNames() != null) {
+          String firstTableName = null;
+          for (String tableName : op.getTableNames()) {
+            TagsSet tagsSet = TagsSet.of(
+                dbTag.getKey(), dbTag.getValue(),
+                TwContextMetricsTemplate.TAG_EP_GROUP, context.getGroup(),
+                TwContextMetricsTemplate.TAG_EP_NAME, context.getName(),
+                TwContextMetricsTemplate.TAG_EP_OWNER, context.getOwner(),
+                inTransactionTag.getKey(), inTransactionTag.getValue(),
+                TAG_OPERATION, opName,
+                successTag.getKey(), successTag.getValue(),
+                TAG_TABLE, tableName);
 
-          if (firstTableName == null) {
-            firstTableName = tableName;
-            meterCache.timer(TIMER_FIRST_TABLE_ACCESS, tagsSet).record(executionTimeNs, TimeUnit.NANOSECONDS);
+            if (firstTableName == null) {
+              firstTableName = tableName;
+              meterCache.timer(TIMER_FIRST_TABLE_ACCESS, tagsSet).record(executionTimeNs, TimeUnit.NANOSECONDS);
+            }
+            meterCache.counter(COUNTER_TABLE_ACCESS, tagsSet).increment();
           }
-          meterCache.counter(COUNTER_TABLE_ACCESS, tagsSet).increment();
         }
       }
-    }
-  }
-
-  @Data
-  @Accessors(chain = true)
-  static class SqlParseResult {
-
-    @NonNull
-    private Map<String, SqlOperation> operations = new HashMap<>();
-
-    @Data
-    @Accessors(chain = true)
-    static class SqlOperation {
-
-      private Set<String> tableNames = new HashSet<>();
     }
   }
 }