+++ tw-entrypoints/src/main/java/com/transferwise/common/entrypoints/tableaccessstatistics/TasSqlFilter.java
@@ -1,0 +1,6 @@
+package com.transferwise.common.entrypoints.tableaccessstatistics;
+
+public interface TasSqlFilter {
+
+  boolean skip(String sql);
+}